Fri, Apr 5, 2024, 1:40 AM +02:00

+++ cash.c
@@ -10,10 +10,10 @@ int main (void)
         n=get_int("change owned: ");
     }
 
-    float x=(n/25);
-    float y=(n/10 - x*2.5);
-    float z=(n/5 - x*5);
-    float v=(n/1 - x*25);
+    int x=(n/25);
+    int y=(n/10 - x*2.5);
+    int z=(n/5 - x*5);
+    int v=(n/1 - x*25);
     int i=(x+y+z+v);
     printf("%i\n",i);
